More efficient binary-fission crafting
When chopping a bell, split it in half, like
the way lenses split, at the nearest halftone
boundary.  Stacking non-full-height bells on
top of one another sticks them back
together to make longer ones.  This
reduces the number of steps necessary to
get to many notes, and increases the value
of planning ahead (e.g. maintaining a
collection of power-of-2-length bells).