Use hammers to ring bells instead of chisels (#5)
Ring bells with hammers, adjusting volume depending on the level of the hammer. The player can punch a bell with a hammer, a door can apply a hammer to a bell, and a door can hit a bell directly with volume depending on the door's level, although this requires a backstop.